Redownload partial music files
When interrupting a download process, only a part of the current file being downloading will be written to the disk. This results in a corrupted .mp3 file. Even redownloading won't fix the file content, since the file with that name already exists.

This change adds an additional check for the file size, to ensure the file to be redownloaded if it's not complete.